+++ Facade_Engine/Compute/AdjacentElements.cs
@@ -42,10 +42,10 @@ namespace BH.Engine.Facade
         /****          Public Methods                   ****/
         /***************************************************/
 
-        [Description("Returns elements adjacent to a provided Element2D from a collection of elements")]
-        [Input("elem", "Element to find adjacencies at")]
-        [Input("refElems", "Elements to use to find element adjacencies")]
-        [Output("adjElems", "Adjacents elements to the provided element")]
+        [Description("Returns elements adjacent to a provided Element2D from a collection of elements.")]
+        [Input("elem", "Element to find adjacencies at.")]
+        [Input("refElems", "Elements to use to find element adjacencies.")]
+        [Output("adjElems", "Adjacents elements to the provided element.")]
         public static List<IElement2D> AdjacentElements(this IElement2D elem, List<IElement2D> refElems)
         {
             List<IElement2D> adjElems = new List<IElement2D>();
